GBy HashMap BM: Separate helpers from test
